More changes to the interface to add more polish.
Fixed a bug where all captions in PowerPoints were being placed at the bottom of the slide

Fixed a bug that occured when captions appeared within an image element of a word document. Word images can now have a caption associated with them if the caption occurs in the same <w:p> tag as the image.